Backend/requirements.txt
Flask==3.0.0
Flask-Cors
python-dotenv
Werkzeug
langchain
langchain-google-genai